fix: Shortcuts don't register when opening a file directly.
When opening Blender by double-clicking on a file, the Gizmodal Ops Shortcuts didn't register. This was due to the order in which Blender registers addons when opening a file: Start Blender -> Register addons (restricted context) -> Blender is open, the context is still restricted -> Load the file. This last step wipes out all non-persistent app.handlers and app.timers which lead to the register_keymap function never being run in unrestriced context.