fix: Avoid producing an error when trying to unregister the register_keymap timer.
When trying to disable Gizmodal Ops, an error message was popping up. While everything worked as expected, seeing an error message is very inconvenient. The error was caused because most of the time, the register_keymap timer is already unregistered. Therefore, this attempt of unregistering is now wrapped inside a try-except block and prints the error message to the console, so debugging is still possible.