fix(Operators): Hardcoded keys in self.key lead to accidental activation of the modal operators.
This commit changes the behavior of Gizmodal Ops regarding when the TIME_WINDOW is activated. Instead of waiting for the same key to be released, every keyup event counts as release of the original key. Keypresses are not passed through, so no other operators are activated accidentally while holding the original key down.

Closes #2 and #4